# Labor.API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Labor.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Admin/labor/{laborId}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "laborId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AdminUpdateLaborRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/api/Labor/search": {
      "get": {
        "tags": [
          "Labor"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "availabilityDate",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "LaborTypeId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "SearchText",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Latitude",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "Longitude",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "RadiusKm",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "MinRating",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "MaxDailyRate",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "AvailabilityStatus",
            "type": "string"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LaborResponseDtoIEnumerableApiResponse"
            }
          }
        }
      }
    },
    "/api/Labor/nearby": {
      "get": {
        "tags": [
          "Labor"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "latitude",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "longitude",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "radiusKm",
            "type": "integer",
            "format": "int32",
            "default": 10
          },
          {
            "in": "query",
            "name": "laborTypeId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LaborResponseDtoIEnumerableApiResponse"
            }
          }
        }
      }
    },
    "/api/Labor/nearby-tomorrow": {
      "get": {
        "tags": [
          "Labor"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "latitude",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "longitude",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "radiusKm",
            "type": "integer",
            "format": "int32",
            "default": 10
          },
          {
            "in": "query",
            "name": "availabilityStatus",
            "type": "string",
            "default": "Available"
          },
          {
            "in": "query",
            "name": "pageNumber",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 50
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LaborResponseDtoIEnumerableApiResponse"
            }
          }
        }
      }
    },
    "/api/Labor/{laborId}": {
      "get": {
        "tags": [
          "Labor"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "laborId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LaborResponseDtoApiResponse"
            }
          }
        }
      }
    },
...
    "/api/Labor/{laborId}/availability-status": {
      "put": {
        "tags": [
          "Labor"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "laborId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateAvailabilityDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/api/Labor/{laborId}/skills": {
      "post": {
        "tags": [
          "Labor"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "laborId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LaborSkillDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/api/Labor/skills/{skillId}": {
      "delete": {
        "tags": [
          "Labor"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "skillId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/api/Labor/{laborId}/availabilities": {
      "put": {
        "tags": [
          "Labor"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "laborId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpsertLaborAvailabilityRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Labor"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "laborId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "year",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "month",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LaborAvailabilityItemDtoIEnumerableApiResponse"
            }
          }
        }
      }
    },
...
    "/api/Review/labor/{laborId}": {
      "get": {
        "tags": [
          "Review"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "laborId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageNumber",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LaborReviewDtoIEnumerableApiResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AdminUpdateLaborRequestDto": {
      "required": [
        "city",
        "country",
        "firstName",
        "laborTypeId",
        "lastName",
        "mobileNumber",
        "state",
        "street",
        "zipCode"
      ],
      "type": "object",
      "properties": {
        "mobileNumber": {
          "maxLength": 15,
          "minLength": 0,
          "type": "string"
        },
        "userName": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "firstName": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "lastName": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "email": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "profilePicture": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "newPassword": {
          "minLength": 6,
          "type": "string"
        },
        "laborTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "specialization": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "experienceYears": {
          "format": "int32",
          "maximum": 80,
          "minimum": 0,
          "type": "integer"
        },
        "dailyRate": {
          "format": "double",
          "maximum": 79228162514264337593543950335,
          "minimum": 0,
          "type": "number"
        },
        "minimumHours": {
          "format": "int32",
          "maximum": 168,
          "minimum": 1,
          "type": "integer"
        },
        "maximumHours": {
          "format": "int32",
          "maximum": 168,
          "minimum": 1,
          "type": "integer"
        },
        "availabilityStatus": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "street": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "city": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "state": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "country": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "zipCode": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "latitude": {
          "format": "double",
          "type": "number"
        },
        "longitude": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
...
    "ApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": { },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "LaborAvailabilityItemDto": {
      "type": "object",
      "properties": {
        "availableDate": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "startTime": {
          "format": "date-span",
          "type": "string"
        },
        "endTime": {
          "format": "date-span",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "LaborAvailabilityItemDtoIEnumerableApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LaborAvailabilityItemDto"
          }
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "LaborResponseDto": {
      "type": "object",
      "properties": {
        "laborId": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "profilePicture": {
          "type": "string"
        },
        "laborTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "laborType": {
          "type": "string"
        },
        "specialization": {
          "type": "string"
        },
        "experienceYears": {
          "format": "int32",
          "type": "integer"
        },
        "rating": {
          "format": "double",
          "type": "number"
        },
        "totalReviews": {
          "format": "int32",
          "type": "integer"
        },
        "dailyRate": {
          "format": "double",
          "type": "number"
        },
        "minimumHours": {
          "format": "int32",
          "type": "integer"
        },
        "maximumHours": {
          "format": "int32",
          "type": "integer"
        },
        "availabilityStatus": {
          "type": "string"
        },
        "isVerified": {
          "type": "boolean"
        },
        "distanceKm": {
          "format": "double",
          "type": "number"
        },
        "city": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "availableDate": {
          "format": "date-time",
          "type": "string"
        },
        "startTime": {
          "type": "string"
        },
        "endTime": {
          "type": "string"
        },
        "skills": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LaborSkillDto"
          }
        },
        "recentReviews": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LaborReviewDto"
          }
        }
      },
      "additionalProperties": false
    },
    "LaborResponseDtoApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/LaborResponseDto"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
    "LaborResponseDtoIEnumerableApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LaborResponseDto"
          }
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "LaborReviewDto": {
      "required": [
        "rating"
      ],
      "type": "object",
      "properties": {
        "reviewId": {
          "format": "int32",
          "type": "integer"
        },
        "orderItemId": {
          "format": "int32",
          "type": "integer"
        },
        "rating": {
          "format": "int32",
          "maximum": 5,
          "minimum": 1,
          "type": "integer"
        },
        "comment": {
          "maxLength": 1000,
          "minLength": 0,
          "type": "string"
        },
        "employerName": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "LaborReviewDtoIEnumerableApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LaborReviewDto"
          }
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "LaborSkillDto": {
      "required": [
        "proficiencyLevel",
        "skillName"
      ],
      "type": "object",
      "properties": {
        "skillId": {
          "format": "int32",
          "type": "integer"
        },
        "skillName": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "proficiencyLevel": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateAvailabilityDto": {
      "type": "object",
      "properties": {
        "availabilityStatus": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UpsertLaborAvailabilityRequestDto": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LaborAvailabilityItemDto"
          }
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}